requirements.txt
pandas==1.3.3
openpyxl==3.0.9
minio==7.1.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
